requirements.txt
asgiref==3.8.1 ; python_version >= "3.12" and python_version < "4.0"
django-cors-headers==4.7.0 ; python_version >= "3.12" and python_version < "4.0"
django-filter==25.1 ; python_version >= "3.12" and python_version < "4.0"
django-rest-framework==0.1.0 ; python_version >= "3.12" and python_version < "4.0"
django==5.2 ; python_version >= "3.12" and python_version < "4.0"
djangorestframework-simplejwt==5.5.0 ; python_version >= "3.12" and python_version < "4.0"
djangorestframework==3.16.0 ; python_version >= "3.12" and python_version < "4.0"
numpy==2.2.5 ; python_version >= "3.12" and python_version < "4.0"
pandas==2.2.3 ; python_version >= "3.12" and python_version < "4.0"
pyjwt==2.9.0 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
pytz==2025.2 ; python_version >= "3.12" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.12" and python_version < "4.0"
sqlparse==0.5.3 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2025.2 ; python_version >= "3.12" and python_version < "4.0"
